Make WiFi driver module definition conditional
Add a new WIFI_DRIVER_HAL_PERIPHERAL variable that is used to
select the wcn3620 module.  This way, if more than one peripheral
is possible for $(soc_name), the system builder can select from
the various packages that can can provide WIFI_DRIVER_HAL_MODULE
based on the peripheral that is included.

BUG: 24373777
Change-Id: Ibdf9b3fb9c2f6775275a001e436cdbfaee1cb208
TEST: Recompile for dragon board, ensure softap0 or wlan0 appears